Renamed data/trait to temp names
