+++ src/index.mjs
@@ -1,5 +1,5 @@
-export { Data, isData, isSingleton, variant, variantName } from './Data.mjs';
-export { Trait, apply, data, isTrait, _ } from './Trait.mjs';
+export { Data, isData, isSingleton, variant, variantName } from './data2.mjs';
+export { Trait, apply, data, isTrait, _ } from './trait2.mjs';
 export { memoFix } from './memoFix.mjs';
 
 /**
