+++ src/trait2.mjs
